Renamed annotated_state folder to state and updated AnnotatedState repr with full path
